Allow comma as decimal separator in currency input
